+++ editor/svg-editor.js
@@ -180,7 +180,8 @@ const callbacks = [],
   * @property {boolean} [emptyStorageOnDecline=false] Used by `ext-storage.js`; empty any prior storage if the user declines to store
   * @property {string[]} [extensions=module:SVGEditor~defaultExtensions] Extensions to load on startup. Use an array in `setConfig` and comma separated file names in the URL. Extension names must begin with "ext-". Note that as of version 2.7, paths containing "/", "\", or ":", are disallowed for security reasons. Although previous versions of this list would entirely override the default list, as of version 2.7, the defaults will always be added to this explicit list unless the configuration `noDefaultExtensions` is included.
   * @property {module:SVGEditor.Stylesheet[]} [stylesheets=["@default"]] An array of required stylesheets to load in parallel; include the value `"@default"` within this array to ensure all default stylesheets are loaded.
-  * @property {string[]} [allowedOrigins=[]] Used by `ext-xdomain-messaging.js` to indicate which origins are permitted for cross-domain messaging (e.g., between the embedded editor and main editor code). Besides explicit domains, one might add '' to allow all domains (not recommended for privacy/data integrity of your user's content!), `window.location.origin` for allowing the same origin (should be safe if you trust all apps on your domain), 'null' to allow `file://` URL usage
+  * @property {string[]} [allowedOrigins=[]] Used by `ext-xdomain-messaging.js` to indicate which origins are permitted for cross-domain messaging (e.g., between the embedded editor and main editor code). Besides explicit domains, one might add '*' to allow all domains (not recommended for privacy/data integrity of your user's content!), `window.location.origin` for allowing the same origin (should be safe if you trust all apps on your domain), 'null' to allow `file:///` URL usage
+  * @property {string[]} [allowedImageLibOrigins=[]] Used by `ext-imagelib.js` to indicate which origins are permitted for cross-domain image loading (image libraries). Besides explicit domains, one might add '*' to allow all domains (not recommended for privacy/data integrity of your user's content!), `window.location.origin` for allowing the same origin (should be safe if you trust all apps on your domain), 'null' to allow `file:///` URL usage
   * @property {null|PlainObject} [colorPickerCSS=null] Object of CSS properties mapped to values (for jQuery) to apply to the color picker. See {@link http://api.jquery.com/css/#css-properties}. A `null` value (the default) will cause the CSS to default to `left` with a position equal to that of the `fill_color` or `stroke_color` element minus 140, and a `bottom` equal to 40
   * @property {string} [paramurl] This was available via URL only. Allowed an un-encoded URL within the query string (use "url" or "source" with a data: URI instead)
   * @property {Float} [canvas_expansion=3] The minimum area visible outside the canvas, as a multiple of the image dimensions. The larger the number, the more one can scroll outside the canvas.
@@ -293,20 +294,22 @@ let svgCanvas, urldata,
     extensions: [],
     stylesheets: [],
     /**
-    * Can use window.location.origin to indicate the current
+    * Can use `location.origin` to indicate the current
     * origin. Can contain a '*' to allow all domains or 'null' (as
-    * a string) to support all file:// URLs. Cannot be set by
+    * a string) to support all `file:///` URLs. Cannot be set by
     * URL for security reasons (not safe, at least for
     * privacy or data integrity of SVG content).
     * Might have been fairly safe to allow
-    *   `new URL(window.location.href).origin` by default but
+    *   `new URL(location.href).origin` by default but
     *   avoiding it ensures some more security that even third
     *   party apps on the same domain also cannot communicate
     *   with this app by default.
-    * For use with ext-xdomain-messaging.js
+    * `allowedOrigins` is for use with `ext-xdomain-messaging.js` and
+    * `allowedImageLibOrigins` is for use with `ext-imagelib.js`
     * @todo We might instead make as a user-facing preference.
     */
-    allowedOrigins: []
+    allowedOrigins: [],
+    allowedImageLibOrigins: []
   };
 
 function loadSvgString (str, callback) {
@@ -487,11 +490,11 @@ editor.setConfig = function (opts, cfgCfg) {
         } else {
           $.pref(key, val);
         }
-      } else if (['extensions', 'stylesheets', 'allowedOrigins'].includes(key)) {
+      } else if (['extensions', 'stylesheets', 'allowedOrigins', 'allowedImageLibOrigins'].includes(key)) {
         if (cfgCfg.overwrite === false &&
           (
             curConfig.preventAllURLConfig ||
-            ['allowedOrigins', 'stylesheets'].includes(key) ||
+            ['allowedOrigins', 'allowedImageLibOrigins', 'stylesheets'].includes(key) ||
             (key === 'extensions' && curConfig.lockExtensions)
           )
         ) {
@@ -671,7 +674,7 @@ editor.init = function () {
       curConfig.extensions = curConfig.extensions.concat(defaultExtensions);
     }
     // ...and remove any dupes
-    ['extensions', 'stylesheets', 'allowedOrigins'].forEach(function (cfg) {
+    ['extensions', 'stylesheets', 'allowedOrigins', 'allowedImageLibOrigins'].forEach(function (cfg) {
       curConfig[cfg] = $.grep(curConfig[cfg], function (n, i) { // Supposedly faster than filter per http://amandeep1986.blogspot.hk/2015/02/jquery-grep-vs-js-filter.html
         return i === curConfig[cfg].indexOf(n);
       });