+++ editor/xdomain-svgedit-config-es.js
@@ -1,4 +1,6 @@
 import svgEditor from './svg-editor.js';
 svgEditor.setConfig({
-  allowedOrigins: ['*']
+  canvasName: 'xdomain', // Namespace this
+  allowedOrigins: ['*'],
+  allowedImageLibOrigins: ['*']
 });